+++ lib/glimmer/swing/component_listener_proxy.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2021 Andy Maleh
+# Copyright (c) 2007-2021 Andy Maleh
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -19,17 +19,30 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/dsl/static_expression'
-require 'glimmer/swing/component_proxy'
-
 module Glimmer
-  module DSL
-    module Swing
-      class OnExpression < StaticExpression
-        def interpret(parent, keyword, *args, &block)
-          # TODO
-        end
+  module Swing
+    # Proxy for component listeners
+    #
+    # Follows the Proxy Design Pattern
+    class ComponentListenerProxy
+      attr_reader :component, :listener, :component_add_listener_method, :listener_class, :listener_method
+
+      def initialize(component: nil, listener:, component_add_listener_method: nil, listener_class: nil, listener_method: nil)
+        @component = component
+        @listener = listener
+        @component_add_listener_method = component_add_listener_method
+        @listener_class = listener_class
+        @listener_method = listener_method
       end
+      
+      def component_remove_listener_method
+        @component_add_listener_method.sub('add', 'remove')
+      end
+      
+      def deregister
+        @component.send(component_remove_listener_method, @listener)
+      end
+      alias unregister deregister # TODO consider dropping unregister (and in Observer too)
     end
   end
 end