+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 - Initial version of Glimmer DSL for Swing
 - General component support
 - General component property support
+- General component listener support
 - Hello, World! sample
+- Hello, Button! sample